ID: 3 #TestCase Update : Refer note section of test case
1.The code to setup and launch WebDriver (Firefox) should be improved to
include Location   Firefox.exe    and a Firefox Profile must be created.
2.The amount of time the driver should wait when searching for a GUI
element should be specified
The code to setup and launch Firefox should be created as a separate
method.
3.All script initialization parameters like Location of Firefox , Base
URL, User Name , Password etc should be stored in separate file say
Util.java.  This helps in easy maintenance of script